+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Top-level `Version()` and `SemVersion()` functions defining the current version of the contrib package.
+- Top-level `Version()` and `SemVersion()` functions defining the current version of the contrib package. (#225)
 
 ## [0.10.1] - 2020-08-13